Simplification of code.
Author entity no longer exists: instead there is an author list string
on the Book entity. This simplifies things in quite a few places.
Additionally, the callbacks on the Search API clients no longer deal
with the BookMetadata holding bays, instead receive the JSON response,
and it is up to the callbacks to parse them into Book objects.